calculate

+++ AreaOfTriangle.java
@@ -5,9 +5,9 @@ public class AreaOfTriangle{
     public static void main(String[] args) {
         Scanner input=new Scanner(System.in);
         System.out.println("Enter a base=");
-        int b=input.nextInt();
+        float b=input.nextFloat();
         System.out.println("Enter a height=");
-        int h=input.nextInt();
+        float h=input.nextFloat();
         float AT=(b*h)/2;
         System.out.println("area of triangle is="+AT);
     }
